requirements.txt
Django==3.0.9
djangorestframework==3.11.0

# Testing
pytest==6.0.1
flake8==3.8.3
flake8-django==1.1.1

# Development
django_extensions==3.0.3